+++ .github/scripts/daily_email_team_flask.py
@@ -1,6 +1,7 @@
-from core.managers.config_manager import ConfigManager
 from flask import Flask
 from flask_mail import Mail, Message
+
+from core.managers.config_manager import ConfigManager
 
 # Crear app y cargar configuración
 app = Flask(__name__)