+++ app/modules/comments/tests/test_selenium.py
@@ -1,16 +1,15 @@
 import time
-from selenium.webdriver.common.by import By
+
 from selenium.common.exceptions import NoSuchElementException
+from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 
 from core.environment.host import get_host_for_selenium_testing
-from core.selenium.common import initialize_driver, close_driver
+from core.selenium.common import close_driver, initialize_driver
 
 
 def click_safely(driver, by, selector):
-    elem = WebDriverWait(driver, 10).until(
-        lambda d: d.find_element(by, selector)
-    )
+    elem = WebDriverWait(driver, 10).until(lambda d: d.find_element(by, selector))
     driver.execute_script("arguments[0].scrollIntoView({block: 'center'});", elem)
     time.sleep(0.2)
     elem.click()
@@ -31,18 +30,14 @@ def test_comments_flow():
         driver.find_element(By.ID, "password").send_keys("1234")
         click_safely(driver, By.ID, "submit")
 
+        driver.find_element(By.LINK_TEXT, "Explore").click()
         click_safely(driver, By.LINK_TEXT, "Weather Data (V2)")
 
         driver.find_element(By.ID, "content").send_keys("esto es una prueba")
         click_safely(driver, By.CSS_SELECTOR, "form > .btn")
 
-        click_safely(driver, By.LINK_TEXT, "Home")
-
-        click_safely(driver, By.CSS_SELECTOR, ".card:nth-child(4) .d-flex")
+        driver.find_element(By.LINK_TEXT, "Explore").click()
         click_safely(driver, By.LINK_TEXT, "UVL Models (V1)")
-        # Para esta última parte, asegurarse de que el comentario no esté aprobado en la base de datos
-        # si lo está, el test fallará al no encontrar el comentario, y debe cambiar el 2, por otro índice
-        #
         click_safely(driver, By.XPATH, "//button[contains(text(), 'Approve') or contains(@class, 'btn-success')]")
 
         time.sleep(2)